+++ src/components/new-landing-page/contributors/info.tsx
