+++ src/components/new-landing-page/header.tsx
@@ -1,6 +1,8 @@
 import React from 'react'
 import { Link } from 'gatsby'
-import RoarLogo from './roar-logo'
+import MHIDotsLogo from '../shared/mhi-dots-logo'
+
+// import { MHIRotatingLogo } from './rotating-logo'
 
 type HeaderProps = {
   headerRef?: React.MutableRefObject<any>
@@ -11,8 +13,8 @@ type HeaderProps = {
 export default function Header({ headerRef, heroRef, otherLinks }: HeaderProps): JSX.Element {
   return (
     <header className="layout-new-header new-landing-page" ref={headerRef as any}>
-      <Link ref={heroRef} className="roar-home home active" to="/roar" aria-label="logo">
-        <RoarLogo />
+      <Link ref={heroRef} className="roar-home home active" to="/" aria-label="logo">
+        <MHIDotsLogo />
       </Link>
       {otherLinks && <div className="other-links">{otherLinks}</div>}
     </header>
