+++ src/components/new-landing-page/hero/dots.tsx
