+++ src/components/new-landing-page/hero/index.tsx
