+++ src/components/new-landing-page/index.tsx
@@ -1,14 +1,14 @@
 import React from 'react'
-import { Layout } from '../../shared/layout'
+import { Layout } from '../shared/layout'
 import useHeader from './useHeader'
 
-import GetUpdates from '../../shared/get-updates'
+import GetUpdates from '../shared/get-updates'
 import Hero from './hero'
 import Causes from './causes'
 import Contributors from './contributors'
 import SocialMediaBar from './social-media-bar'
-import useExplicitHeightOnIPhone from '../../../effects/useExplicitHeightOnIPhone'
-import RoarSEO from '../roar-seo'
+import useExplicitHeightOnIPhone from '../../effects/useExplicitHeightOnIPhone'
+import SEO from '../shared/seo'
 
 const NewLandingPage = ({ location }: PageProps): JSX.Element => {
   const navigator = typeof window === 'undefined' ? undefined : window.navigator
@@ -25,7 +25,7 @@ const NewLandingPage = ({ location }: PageProps): JSX.Element => {
         <div style={{ width: '100%', height: '200px', backgroundColor: 'red' }} />
       }
     >
-      <RoarSEO />
+      <SEO />
       <Hero ref={internalSectionRefs['hero']} dotsRef={dotsRef} />
       <Causes ref={internalSectionRefs['Causes']} />
       <Contributors ref={internalSectionRefs['Contributors']} />
