--- src/components/new-landing-page/roar-logo.tsx
@@ -1,44 +1,0 @@
-import React from 'react'
-
-export default (): JSX.Element => {
-  return (
-    <svg className="roar-logo" viewBox="0 0 164 60" fill="none" xmlns="http://www.w3.org/2000/svg">
-      <g id="Roar! Logo">
-        <path
-          id="!"
-          d="M154.766 6.14404V24.448C154.766 25.408 154.745 26.3467 154.702 27.264C154.66 28.1814 154.606 29.1094 154.542 30.048C154.478 30.9654 154.393 31.9147 154.286 32.896C154.201 33.856 154.094 34.8907 153.966 36H150.126C149.998 34.8907 149.881 33.856 149.774 32.896C149.689 31.9147 149.614 30.9654 149.55 30.048C149.486 29.1094 149.433 28.1814 149.39 27.264C149.348 26.3467 149.326 25.408 149.326 24.448V6.14404H154.766ZM147.95 48.48C147.95 47.9254 148.046 47.4027 148.238 46.912C148.452 46.4214 148.74 45.9947 149.102 45.632C149.465 45.2694 149.881 44.9814 150.35 44.768C150.841 44.5547 151.374 44.448 151.95 44.448C152.505 44.448 153.017 44.5547 153.486 44.768C153.977 44.9814 154.404 45.2694 154.766 45.632C155.129 45.9947 155.417 46.4214 155.63 46.912C155.844 47.4027 155.95 47.9254 155.95 48.48C155.95 49.056 155.844 49.5894 155.63 50.08C155.417 50.5494 155.129 50.9654 154.766 51.328C154.404 51.6907 153.977 51.968 153.486 52.16C153.017 52.3734 152.505 52.48 151.95 52.48C151.374 52.48 150.841 52.3734 150.35 52.16C149.881 51.968 149.465 51.6907 149.102 51.328C148.74 50.9654 148.452 50.5494 148.238 50.08C148.046 49.5894 147.95 49.056 147.95 48.48Z"
-          fill="#FA759E"
-        />
-        <path
-          id="ar"
-          d="M126.849 25.8117C127.893 23.57 129.177 21.8242 130.699 20.574C132.222 19.3023 134.083 18.6665 136.28 18.6665C136.976 18.6665 137.64 18.7419 138.271 18.8928C138.923 19.0437 139.5 19.2808 140 19.6041L139.576 23.9041C139.445 24.443 139.119 24.7124 138.597 24.7124C138.292 24.7124 137.846 24.6478 137.259 24.5184C136.671 24.3891 136.008 24.3244 135.268 24.3244C134.224 24.3244 133.288 24.4753 132.462 24.7771C131.657 25.0788 130.928 25.5315 130.275 26.135C129.644 26.7169 129.068 27.4498 128.546 28.3335C128.045 29.1957 127.588 30.1871 127.175 31.308V51.9998H121.333V19.2485H124.662C125.293 19.2485 125.728 19.367 125.968 19.6041C126.207 19.8412 126.37 20.2507 126.457 20.8327L126.849 25.8117Z"
-          fill="#164176"
-        />
-        <path
-          id="a"
-          d="M106.08 37.4078C103.456 37.4931 101.216 37.7065 99.3601 38.0478C97.5254 38.3678 96.0214 38.7945 94.8481 39.3278C93.6961 39.8611 92.8534 40.4905 92.3201 41.2158C91.8081 41.9411 91.5521 42.7518 91.5521 43.6478C91.5521 44.5011 91.6907 45.2371 91.9681 45.8558C92.2454 46.4745 92.6187 46.9865 93.0881 47.3918C93.5787 47.7758 94.1441 48.0638 94.7841 48.2558C95.4454 48.4265 96.1494 48.5118 96.8961 48.5118C97.8987 48.5118 98.8161 48.4158 99.6481 48.2238C100.48 48.0105 101.259 47.7118 101.984 47.3278C102.731 46.9438 103.435 46.4851 104.096 45.9518C104.779 45.4185 105.44 44.8105 106.08 44.1278V37.4078ZM87.6801 24.1598C89.4721 22.4318 91.4027 21.1411 93.4721 20.2878C95.5414 19.4345 97.8347 19.0078 100.352 19.0078C102.165 19.0078 103.776 19.3065 105.184 19.9038C106.592 20.5011 107.776 21.3331 108.736 22.3998C109.696 23.4665 110.421 24.7571 110.912 26.2718C111.403 27.7865 111.648 29.4505 111.648 31.2638V51.9998H109.12C108.565 51.9998 108.139 51.9145 107.84 51.7438C107.541 51.5518 107.307 51.1891 107.136 50.6558L106.496 47.5838C105.643 48.3731 104.811 49.0772 104 49.6958C103.189 50.2931 102.336 50.8052 101.44 51.2318C100.544 51.6371 99.5841 51.9465 98.5601 52.1598C97.5574 52.3945 96.4374 52.5118 95.2001 52.5118C93.9414 52.5118 92.7574 52.3411 91.6481 51.9998C90.5387 51.6371 89.5681 51.1038 88.7361 50.3998C87.9254 49.6958 87.2747 48.8105 86.7841 47.7438C86.3147 46.6558 86.0801 45.3758 86.0801 43.9038C86.0801 42.6238 86.4321 41.3971 87.1361 40.2238C87.8401 39.0291 88.9814 37.9731 90.5601 37.0558C92.1387 36.1385 94.1974 35.3918 96.7361 34.8158C99.2747 34.2185 102.389 33.8771 106.08 33.7918V31.2638C106.08 28.7465 105.536 26.8478 104.448 25.5678C103.36 24.2665 101.771 23.6158 99.6801 23.6158C98.2721 23.6158 97.0881 23.7971 96.1281 24.1598C95.1894 24.5011 94.3681 24.8958 93.6641 25.3438C92.9814 25.7705 92.3841 26.1651 91.8721 26.5278C91.3814 26.8691 90.8907 27.0398 90.4001 27.0398C90.0161 27.0398 89.6854 26.9438 89.4081 26.7518C89.1307 26.5385 88.8961 26.2825 88.7041 25.9838L87.6801 24.1598Z"
-          fill="#164176"
-        />
-        <g id="LOGO">
-          <path
-            id="Ellipse 3"
-            d="M80 28.0471C80 41.3223 72.8366 52.0001 64 52.0001C55.1635 52.0001 48 41.3223 48 28.0471C48 14.772 55.1635 21.9238 64 21.9238C72.8366 21.9238 80 14.772 80 28.0471Z"
-            fill="#164176"
-          />
-          <path
-            id="Intersect"
-            fillRule="evenodd"
-            clipRule="evenodd"
-            d="M52.0132 43.9318C52.6729 39.4261 57.792 35.9199 64.0091 35.9199C70.2158 35.9199 75.3281 39.4144 76.0017 43.9091C73.0699 48.8777 68.78 52.0012 64.0007 52.0012C59.2287 52.0012 54.9447 48.8872 52.0132 43.9318Z"
-            fill="#FA759E"
-          />
-        </g>
-        <path
-          id="r"
-          d="M23.4975 28.352C25.2895 28.352 26.8681 28.1387 28.2335 27.712C29.6201 27.264 30.7721 26.6454 31.6895 25.856C32.6281 25.0454 33.3321 24.0854 33.8015 22.976C34.2708 21.8667 34.5055 20.64 34.5055 19.296C34.5055 16.5654 33.6095 14.5067 31.8175 13.12C30.0255 11.7334 27.3375 11.04 23.7535 11.04H16.9695V28.352H23.4975ZM44.6815 52H39.1775C38.0468 52 37.2148 51.5627 36.6815 50.688L24.7775 34.304C24.4148 33.792 24.0201 33.4294 23.5935 33.216C23.1881 32.9814 22.5481 32.864 21.6735 32.864H16.9695V52H10.7935V6.14404H23.7535C26.6548 6.14404 29.1615 6.44271 31.2735 7.04004C33.3855 7.61604 35.1241 8.45871 36.4895 9.56804C37.8761 10.6774 38.9001 12.0214 39.5615 13.6C40.2228 15.1574 40.5535 16.9067 40.5535 18.848C40.5535 20.4694 40.2975 21.984 39.7855 23.392C39.2735 24.8 38.5268 26.0694 37.5455 27.2C36.5855 28.3094 35.4015 29.2587 33.9935 30.048C32.6068 30.8374 31.0281 31.4347 29.2575 31.84C30.0468 32.288 30.7295 32.9387 31.3055 33.792L44.6815 52Z"
-          fill="#164176"
-        />
-      </g>
-    </svg>
-  )
-}
