+++ src/components/new-landing-page/social-media-bar/email-at-icon.tsx
