+++ src/components/new-landing-page/social-media-bar/index.tsx
