+++ src/components/new-landing-page/social-media-bar/product-hunt-icon.tsx
