+++ src/components/roar/landing-page/how-it-works/index.tsx
